+++ base/strings/strings.jl
@@ -4,7 +4,7 @@ include("strings/substring.jl")
 include("strings/search.jl")
 include("strings/unicode.jl")
 
-import .Unicode: textwidth, islowercase, isuppercase, isalpha, isdigit, isnumeric, iscntrl, ispunct,
+import .Unicode: textwidth, islowercase, isuppercase, isletter, isdigit, isnumeric, iscntrl, ispunct,
     isspace, isprint, isxdigit, lowercase, uppercase, titlecase, lowercasefirst, uppercasefirst
 
 include("strings/util.jl")